homepage changes and stuff

+++ src/main/java/com/group13/fleet/repository/CompanyRepository.java
@@ -8,6 +8,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
 }
